Agregamos manejo de errores en controlador de orderdetail y en dataservices

+++ app/Http/Services/DataServices.php
@@ -11,7 +11,7 @@ class DataServices {
     }
     
   public function getAll(){
-    return $this->model->all();
+    return $this->model->orderBy('id', 'desc')->paginate();
   }
   public function getById($id){
     return $this->model->find($id);
@@ -19,15 +19,24 @@ class DataServices {
   public function create(array $data){
     return $this->model->create($data);
   }
-  public function update($id , array $data){
-    $record = $this->getById($id);
-    $record->update($data);
-    return $record;
+  public function update($id, array $data)
+  {
+      $record = $this->getById($id);
+      if (!$record) {
+          return null;
+      }
+      $record->update($data);
+      return $record;
   }
-  public function delete($id){
-    $record = $this->getById($id);
-    $record->delete();
-    return $record;
+
+  public function delete($id)
+  {
+      $record = $this->getById($id);
+      if (!$record) {
+          return null;
+      }
+      $record->delete();
+      return $record;
   }
